another refactoring.
