+++ src/v2/view-builder/views/idp/BaseIdpAuthenticator.js
@@ -1,15 +1,17 @@
 import { BaseForm } from '../../internals';
+import BaseAuthenticatorView from 'v2/view-builder/components/BaseAuthenticatorView';
 
-export default BaseForm.extend({
-
+export const BaseIdPAuthenticatorBody =  BaseForm.extend({
   initialize() {
     BaseForm.prototype.initialize.apply(this, arguments);
     this.model.set('useRedirect', true);
     this.add('<div class="okta-waiting-spinner"></div>');
-  },
+  }
+});
 
-  postRender: function() {
-    BaseForm.prototype.postRender.apply(this, arguments);
+export const BaseIdpAuthenticatorView = BaseAuthenticatorView.extend({
+  postRender() {
+    BaseAuthenticatorView.prototype.postRender.apply(this, arguments);
     // get the error messages
     const messages = this.options.appState.get('messages') || {};
     // In case of failure, don't auto-redirect which will result in infinite redirects.
@@ -17,7 +19,7 @@ export default BaseForm.extend({
     if (this.settings.get('features.skipIdpFactorVerificationBtn') && !Array.isArray(messages.value)) {
       this.$('.o-form-button-bar').hide();
       this.$('.okta-waiting-spinner').show();
-      this.saveForm(this.model);
+      this.form.trigger('save', this.model);
     } else {
       this.$('.okta-waiting-spinner').hide();
       this.$('.o-form-button-bar').show();
